% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_function.R
\name{collapse.clusters}
\alias{collapse.clusters}
\title{make `dgCMatrix` for a pseudo bulk profiles}
\usage{
\method{collapse}{clusters}(counts, cell.groups)
}
\arguments{
\item{counts}{a sparse matrix instance where the rows are genes
and the columns are cell names}

\item{cell.groups}{a list of cell names with cell type with
annotations}
}
\value{
return the collapsed `dgCMatrix` with gene x cell-types
}
\description{
make `dgCMatrix` for a pseudo bulk profiles
}
