% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_function.R
\name{computeJointContext}
\alias{computeJointContext}
\title{computer joint context}
\usage{
computeJointContext(
  p.obj,
  annotation.type.name = "annot",
  joint.type.name = "jcontext"
)
}
\arguments{
\item{p.obj}{a pagoda object}

\item{annotation.type.name}{name of the annotation object within pagoda object}

\item{joint.type.name}{name of the object that would context specific annotation}
}
\value{
modified pagoda2 object
}
\description{
computer joint context
}
