% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_function.R
\name{construct.subset.profile}
\alias{construct.subset.profile}
\title{construct the profiles within a construct}
\usage{
construct.subset.profile(r, r.raw, cnames, cutoff, special.genes = NULL)
}
\arguments{
\item{r}{the pagoda object}

\item{r.raw}{the SparseSpatial object with raw counts}

\item{cnames}{the cell names}

\item{cutoff}{the cutoff number of cells per gene type}

\item{special.genes}{the special genes that are needed to be included}
}
\value{
profiles
}
\description{
construct the profiles within a construct
}
