% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{convertSpatialObj2Pagoda}
\alias{convertSpatialObj2Pagoda}
\title{Convert the Spatial objects to pagoda
call `createSparseObj` or `createSparseSpatialObjFromRCTD`
before calling this function}
\usage{
convertSpatialObj2Pagoda(cdl)
}
\arguments{
\item{cdl}{a list of SparseSpatialObj objects with names}
}
\value{
a list of containing two objects,
        object 1 is a list of SparseSpatial objects with modified
        names, it contains the raw counts
        object 2 is a list of pagoda2 objects
}
\description{
Convert the Spatial objects to pagoda
call `createSparseObj` or `createSparseSpatialObjFromRCTD`
before calling this function
}
