% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_function.R
\name{createKDEBoundary}
\alias{createKDEBoundary}
\title{create KDE boundary for a pagoda object}
\usage{
createKDEBoundary(p.obj, cell.type = "Tumor")
}
\arguments{
\item{p.obj}{is a pagoda2 object}

\item{cell.type}{a cell type default value is Tumor}
}
\value{
a modified object with boundary stored in
        an object contour misc$cont
        and create primary contexts
}
\description{
create KDE boundary for a pagoda object
}
