% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{createSparseSpatialObj}
\alias{createSparseSpatialObj}
\title{convert the csv file and spatial info}
\usage{
createSparseSpatialObj(gene.matrix.file, location.file)
}
\arguments{
\item{gene.matrix.file}{Direct csv file for gene to cell matrix}

\item{location.file}{file containing the the dataframe}
}
\value{
a list containing a sparse matrix and
        a dataframe
}
\description{
convert the csv file and spatial info
}
