% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{createSparseSpatialObjFromRCTD}
\alias{createSparseSpatialObjFromRCTD}
\title{convert the csv file and spatial info}
\usage{
createSparseSpatialObjFromRCTD(pw)
}
\arguments{
\item{pw}{path received from the RCTD file should
contain the `alignment/MappedDGEForR.csv` which
is a gene to cell count matrix
and another file
`barcode_matching/BeadLocationsForR.csv`
which contains the spatial information}
}
\value{
a list containing a sparse matrix and
        a dataframe
}
\description{
convert the csv file and spatial info
}
