% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_function.R
\name{get.inner.embedding}
\alias{get.inner.embedding}
\title{Get inner embedding from the pagoda2 object}
\usage{
get.inner.embedding(
  r,
  tumor.center,
  nontumor.center,
  ctype = NULL,
  mode = "manual"
)
}
\arguments{
\item{r}{pagoda2 object}

\item{tumor.center}{coordinate for tentative center for the
tumor context}

\item{nontumor.center}{coordinate for tentative center for the
non tumor context}

\item{ctype}{the cell type}

\item{mode}{if we _don't_ fix the centers of the contexts manually
they will be calculated}
}
\value{
a dataframe with the distance from the boundary
}
\description{
Get inner embedding from the pagoda2 object
}
