% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_function.R
\name{get.subset}
\alias{get.subset}
\title{plot the beads with the distance boundary}
\usage{
get.subset(
  r,
  ctype.emb.df,
  distance.from.boundary,
  distance.from.center = NULL,
  pattern = "pattern1"
)
}
\arguments{
\item{r}{a pagoda2 object with all the information}

\item{ctype.emb.df}{the dataframe to be generated from
`get.inner.embedding`}

\item{distance.from.boundary}{the distance from the boundary drawn from KDE}

\item{distance.from.center}{denotes the distance for manually fixed centers
in the respective context regions}

\item{pattern}{TODO: we have to get rid of multiple patterns}
}
\value{
a ggplot object
}
\description{
plot the beads with the distance boundary
}
