% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{get.z.edge.R}
\alias{get.z.edge.R}
\title{get z values from edgeR DE}
\usage{
get.z.edge.R(d)
}
\arguments{
\item{d}{the data.frame}
}
\value{
updated data.frame
}
\description{
get z values from edgeR DE
}
