% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{getDifferentialExpression}
\alias{getDifferentialExpression}
\title{Context-dependent DE approximation using a linear model}
\usage{
getDifferentialExpression(
  profiles,
  target1,
  target2,
  de.method = "edgeR.qlf",
  default.bcv = 0.1,
  gene.scale.function = sqrt,
  library.size.scale = 1e+06,
  remove.zeroes = FALSE
)
}
\arguments{
\item{profiles}{either a matrix of profiles (columns - cell types, rows - genes)
or a list of such matrices, each corresponding to a replicate}

\item{target1}{name of the cell type in context 1 (e.g. Fibroblast_Tumor)}

\item{target2}{name of the cell type in context 2 (e.g. Firboblast_Non_Tumor)}

\item{de.method}{'edgeR' (default) or 'DESeq2'}

\item{default.bcv}{default value of}

\item{gene.scale.function}{function used in library scaling}

\item{library.size.scale}{library size to scale default is 1e6}

\item{remove.zeroes}{a flag to remove zeros}
}
\value{

}
\description{
Context-dependent DE approximation using a linear model
}
