% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_function.R
\name{plotWithDistanceBoundary}
\alias{plotWithDistanceBoundary}
\title{plot the beads with the distance boundary}
\usage{
plotWithDistanceBoundary(
  r,
  ctype.emb.df,
  distance.from.boundary,
  distance.from.center = NULL,
  pattern = "pattern1"
)
}
\arguments{
\item{r}{a pagoda2 object with all the information}

\item{ctype.emb.df}{the dataframe to be generated from
`get.inner.embedding`}

\item{distance.from.boundary}{the distance from the boundary drawn from KDE}

\item{distance.from.center}{the distance from the center of the context region}

\item{pattern}{TODO: we have to get rid of multiple patterns}
}
\value{
a ggplot object
}
\description{
plot the beads with the distance boundary
}
