% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_function.R
\name{preparePseudoBulkProfile}
\alias{preparePseudoBulkProfile}
\title{Prepare aggregated profiles}
\usage{
preparePseudoBulkProfile(
  r,
  r.raw,
  distance.from.boundary,
  tumor.center,
  nontumor.center,
  special.genes = NULL
)
}
\arguments{
\item{r}{the pagoda object}

\item{r.raw}{the SparseSpatial object with raw counts}

\item{distance.from.boundary}{the cell names}

\item{tumor.center}{the coordinates for the tumor center}

\item{nontumor.center}{the coordinates for the nontumor center}

\item{special.genes}{the special genes that are needed to be included}
}
\value{
the collapsed profiles and the dataframe, distance, plots and
        the cell names
}
\description{
Prepare aggregated profiles
}
