% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_function.R
\name{scatterPlotWithCorrection}
\alias{scatterPlotWithCorrection}
\title{Scatter plot with some top DE genes along
with the tumor markers to show the effect of regression}
\usage{
scatterPlotWithCorrection(res, red.marker, target, top.genes)
}
\arguments{
\item{res}{the de object with both the corrected and uncorrected DE results}

\item{red.marker}{marker genes from the cell-type that we control for (tumor)}

\item{target}{the target cell-type to create the title of the plot}

\item{top.genes}{how many DE genes to annotate}
}
\value{
two ggplots appended side by side
}
\description{
Scatter plot with some top DE genes along
with the tumor markers to show the effect of regression
}
