% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{to_sparse}
\alias{to_sparse}
\title{Convert csv file to the sparse matrix}
\usage{
to_sparse(d_table)
}
\arguments{
\item{d_table}{csv file to be read}
}
\value{
a sparse matrix
}
\description{
Convert csv file to the sparse matrix
}
