% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{updateSpatialInfo}
\alias{updateSpatialInfo}
\title{adding the cell type annotation in the pagoda objects
call `convertSpatialObj2Pagoda` before calling this function}
\usage{
updateSpatialInfo(cdl, cdl.p2, anno)
}
\arguments{
\item{cdl}{a list of SparseSpatial}

\item{cdl.p2}{a list of pagoda2 objects made from `convertSpatialObj2Pagoda`}

\item{anno}{an R object containing the annotation file
it assumes a specific format for now and
TODO: we need to change this and make more generic}
}
\value{
a modified pagoda object with the positions and
        spatial positions
}
\description{
adding the cell type annotation in the pagoda objects
call `convertSpatialObj2Pagoda` before calling this function
}
