% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{updateSpatialInfo}
\alias{updateSpatialInfo}
\title{adding the cell type annotation in the pagoda objects
call `convertSpatialObj2Pagoda` before calling this function}
\usage{
updateSpatialInfo(obj, anno)
}
\arguments{
\item{obj}{an object containing the spatial spatial objects}

\item{anno}{an R object containing the annotation file
it assumes a specific format for now and
TODO: we need to change this and make more generic}
}
\value{
a modified pagoda object with the positions and
        spatial positions
}
\description{
adding the cell type annotation in the pagoda objects
call `convertSpatialObj2Pagoda` before calling this function
}
